requirements.txt
flask 
requests
PyJWT
cryptography